+++ scrapyd/webservice.py
@@ -1,19 +1,25 @@
 from __future__ import annotations
 
 import functools
+import json
+import os
 import sys
 import traceback
 import uuid
 import zipfile
 from copy import copy
 from io import BytesIO
+from subprocess import PIPE, Popen
+from typing import ClassVar
 
 from twisted.python import log
-from twisted.web import error, http
+from twisted.web import error, http, resource
 
-from scrapyd.exceptions import EggNotFoundError, ProjectNotFoundError
+from scrapyd.config import Config
+from scrapyd.exceptions import EggNotFoundError, ProjectNotFoundError, RunnerError
 from scrapyd.jobstorage import job_items_url, job_log_url
-from scrapyd.utils import JsonResource, UtilsCache, get_spider_list, native_stringify_dict
+from scrapyd.sqlite import JsonSqliteDict
+from scrapyd.utils import native_stringify_dict
 
 
 def param(
@@ -50,6 +56,93 @@ def param(
         return wrapper
 
     return decorator
+
+
+def get_spider_list(project, runner=None, pythonpath=None, version=None):
+    """Return the spider list from the given project, using the given runner"""
+
+    # UtilsCache uses JsonSqliteDict, which encodes the project's value as JSON, but JSON allows only string keys,
+    # so the stored dict will have a "null" key, instead of a None key.
+    if version is None:
+        version = ""
+
+    if "cache" not in get_spider_list.__dict__:
+        get_spider_list.cache = UtilsCache()
+    try:
+        return get_spider_list.cache[project][version]
+    except KeyError:
+        pass
+
+    if runner is None:
+        runner = Config().get("runner")
+
+    env = os.environ.copy()
+    env["PYTHONIOENCODING"] = "UTF-8"
+    env["SCRAPY_PROJECT"] = project
+    if pythonpath:
+        env["PYTHONPATH"] = pythonpath
+    if version:
+        env["SCRAPYD_EGG_VERSION"] = version
+    pargs = [sys.executable, "-m", runner, "list", "-s", "LOG_STDOUT=0"]
+    proc = Popen(pargs, stdout=PIPE, stderr=PIPE, env=env)
+    out, err = proc.communicate()
+    if proc.returncode:
+        msg = err or out or ""
+        msg = msg.decode("utf8")
+        raise RunnerError(msg)
+
+    spiders = out.decode("utf-8").splitlines()
+    try:
+        project_cache = get_spider_list.cache[project]
+        project_cache[version] = spiders
+    except KeyError:
+        project_cache = {version: spiders}
+    get_spider_list.cache[project] = project_cache
+
+    return spiders
+
+
+class UtilsCache:
+    # array of project name that need to be invalided
+    invalid_cached_projects: ClassVar = []
+
+    def __init__(self):
+        self.cache_manager = JsonSqliteDict(table="utils_cache_manager")
+
+    # Invalid the spider's list's cache of a given project (by name)
+    @staticmethod
+    def invalid_cache(project):
+        UtilsCache.invalid_cached_projects.append(project)
+
+    def __getitem__(self, key):
+        for p in UtilsCache.invalid_cached_projects:
+            if p in self.cache_manager:
+                del self.cache_manager[p]
+        UtilsCache.invalid_cached_projects[:] = []
+        return self.cache_manager[key]
+
+    def __setitem__(self, key, value):
+        self.cache_manager[key] = value
+
+    def __repr__(self):
+        return f"UtilsCache(cache_manager={self.cache_manager!r})"
+
+
+class JsonResource(resource.Resource):
+    json_encoder = json.JSONEncoder()
+
+    def render(self, txrequest):
+        r = resource.Resource.render(self, txrequest)
+        return self.encode_object(r, txrequest)
+
+    def encode_object(self, obj, txrequest):
+        r = "" if obj is None else self.json_encoder.encode(obj) + "\n"
+        txrequest.setHeader("Content-Type", "application/json")
+        txrequest.setHeader("Access-Control-Allow-Origin", "*")
+        txrequest.setHeader("Access-Control-Allow-Methods", "GET, POST, PATCH, PUT, DELETE")
+        txrequest.setHeader("Access-Control-Allow-Headers", " X-Requested-With")
+        txrequest.setHeader("Content-Length", str(len(r)))
+        return r
 
 
 class WsResource(JsonResource):