+++ tests/test_webservice.py
@@ -1,12 +1,20 @@
+import os
+from io import BytesIO
 from pathlib import Path
+from pkgutil import get_data
+from subprocess import Popen
 from unittest import mock
 
 import pytest
+from scrapy.utils.test import get_pythonpath
+from twisted.trial import unittest
 from twisted.web import error
 
-from scrapyd.exceptions import DirectoryTraversalError
+from scrapyd import get_application
+from scrapyd.exceptions import DirectoryTraversalError, RunnerError
 from scrapyd.interfaces import IEggStorage
 from scrapyd.jobstorage import Job
+from scrapyd.webservice import UtilsCache, get_spider_list
 
 
 def fake_list_jobs(*args, **kwargs):
@@ -19,6 +27,97 @@ def fake_list_spiders(*args, **kwargs):
 
 def fake_list_spiders_other(*args, **kwargs):
     return ["quotesbot", "toscrape-css"]
+
+
+def get_pythonpath_scrapyd():
+    scrapyd_path = __import__("scrapyd").__path__[0]
+    return os.path.join(os.path.dirname(scrapyd_path), get_pythonpath(), os.environ.get("PYTHONPATH", ""))
+
+
+class GetSpiderListTest(unittest.TestCase):
+    def setUp(self):
+        path = os.path.abspath(self.mktemp())
+        j = os.path.join
+        eggs_dir = j(path, "eggs")
+        os.makedirs(eggs_dir)
+        dbs_dir = j(path, "dbs")
+        os.makedirs(dbs_dir)
+        logs_dir = j(path, "logs")
+        os.makedirs(logs_dir)
+        os.chdir(path)
+        with open("scrapyd.conf", "w") as f:
+            f.write("[scrapyd]\n")
+            f.write(f"eggs_dir = {eggs_dir}\n")
+            f.write(f"dbs_dir = {dbs_dir}\n")
+            f.write(f"logs_dir = {logs_dir}\n")
+        self.app = get_application()
+
+    def add_test_version(self, file, project, version):
+        eggstorage = self.app.getComponent(IEggStorage)
+        eggfile = BytesIO(get_data("tests", file))
+        eggstorage.put(eggfile, project, version)
+
+    def test_get_spider_list_log_stdout(self):
+        self.add_test_version("logstdout.egg", "logstdout", "logstdout")
+        spiders = get_spider_list("logstdout", pythonpath=get_pythonpath_scrapyd())
+        # If LOG_STDOUT were respected, the output would be [].
+        self.assertEqual(sorted(spiders), ["spider1", "spider2"])
+
+    def test_get_spider_list(self):
+        # mybot.egg has two spiders, spider1 and spider2
+        self.add_test_version("mybot.egg", "mybot", "r1")
+        spiders = get_spider_list("mybot", pythonpath=get_pythonpath_scrapyd())
+        self.assertEqual(sorted(spiders), ["spider1", "spider2"])
+
+        # mybot2.egg has three spiders, spider1, spider2 and spider3...
+        # BUT you won't see it here because it's cached.
+        # Effectivelly it's like if version was never added
+        self.add_test_version("mybot2.egg", "mybot", "r2")
+        spiders = get_spider_list("mybot", pythonpath=get_pythonpath_scrapyd())
+        self.assertEqual(sorted(spiders), ["spider1", "spider2"])
+
+        # Let's invalidate the cache for this project...
+        UtilsCache.invalid_cache("mybot")
+
+        # Now you get the updated list
+        spiders = get_spider_list("mybot", pythonpath=get_pythonpath_scrapyd())
+        self.assertEqual(sorted(spiders), ["spider1", "spider2", "spider3"])
+
+        # Let's re-deploy mybot.egg and clear cache. It now sees 2 spiders
+        self.add_test_version("mybot.egg", "mybot", "r3")
+        UtilsCache.invalid_cache("mybot")
+        spiders = get_spider_list("mybot", pythonpath=get_pythonpath_scrapyd())
+        self.assertEqual(sorted(spiders), ["spider1", "spider2"])
+
+        # And re-deploying the one with three (mybot2.egg) with a version that
+        # isn't the higher, won't change what get_spider_list() returns.
+        self.add_test_version("mybot2.egg", "mybot", "r1a")
+        UtilsCache.invalid_cache("mybot")
+        spiders = get_spider_list("mybot", pythonpath=get_pythonpath_scrapyd())
+        self.assertEqual(sorted(spiders), ["spider1", "spider2"])
+
+    @pytest.mark.skipif(os.name == "nt", reason="get_spider_list() unicode fails on windows")
+    def test_get_spider_list_unicode(self):
+        # mybotunicode.egg has two spiders, araña1 and araña2
+        self.add_test_version("mybotunicode.egg", "mybotunicode", "r1")
+        spiders = get_spider_list("mybotunicode", pythonpath=get_pythonpath_scrapyd())
+
+        self.assertEqual(sorted(spiders), ["araña1", "araña2"])
+
+    def test_failed_spider_list(self):
+        self.add_test_version("mybot3.egg", "mybot3", "r1")
+        pypath = get_pythonpath_scrapyd()
+        # Workaround missing support for context manager in twisted < 15
+
+        # Add -W ignore to sub-python to prevent warnings & tb mixup in stderr
+        def popen_wrapper(*args, **kwargs):
+            cmd, args = args[0], args[1:]
+            cmd = [cmd[0], "-W", "ignore"] + cmd[1:]
+            return Popen(cmd, *args, **kwargs)
+
+        with mock.patch("scrapyd.webservice.Popen", wraps=popen_wrapper):
+            exc = self.assertRaises(RunnerError, get_spider_list, "mybot3", pythonpath=pypath)
+        self.assertRegex(str(exc).rstrip(), r"Exception: This should break the `scrapy list` command$")
 
 
 class TestWebservice: